Create PQSTTest and move there the modified test
